KowihanBackend/spring-boot-acceleo/storage/acceleo-generator/generated/djd_ce45692b/djd/djd/views/hotel_viewset.py
from rest_framework import viewsets, filters
from django_filters.rest_framework import DjangoFilterBackend
from ..models import Hotel
from ..serializers  import HotelSerializer

class HotelViewSet(viewsets.ModelViewSet):
    queryset = Hotel.objects.all()
    serializer_class = HotelSerializer
    filter_backends = [DjangoFilterBackend, filters.SearchFilter, filters.OrderingFilter]
    filterset_fields = '__all__'
    search_fields = ['name', 'address']
    ordering_fields = '__all__'
