KowihanBackend/spring-boot-acceleo/storage/acceleo-generator/generated/djd_ce45692b/djd/requirements.txt
# Django Core
Django==4.2.7  # ✅ VERSION COMPATIBLE
djangorestframework==3.14.0
django-filter==23.3

# API Documentation
drf-spectacular==0.26.5

# CORS & Security
django-cors-headers==4.3.1

# Environment
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.7

# Authentication

# System dependencies
setuptools==68.2.2  # ✅ AJOUTÉ POUR SIMPLEJWT

